Move out parts of `languages.ts` module into separate files
